requirements.txt
langchain==0.1.7
langchain_community==0.0.20
langchain_core==0.1.23
langchain_openai==0.0.6
python-dotenv==1.0.1
streamlit==1.24.1
torch==2.1.2
transformers==4.36.2
